% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{cdpcx_hist}
\alias{cdpcx_hist}
\title{cdpcx_hist
Plots a fancy cdpcx (extended chordal diatonic pitch class plot"}
\usage{
cdpcx_hist(
  data,
  id = NULL,
  colour_chromatic = T,
  percentage = T,
  fill_var = NULL,
  cdpcx_col = "cdpcx_raw"
)
}
\arguments{
\item{data}{(data frame) most contain cdpcx_col as variable with cdpcx data}

\item{id}{(string or integer) Optional set of ids to filter}

\item{colour_chromatic}{(boolean) Flag whether to colour chromatic pitch classes differently}

\item{percentage}{(boolen) Flag whether to use percentage scale on y-axis}

\item{fill_var}{(String) extra fill variable}

\item{cdpcx_col}{(string) column name where cdpcx data is stored, default cdpcx_raw_all}
}
\value{
A ggplot2 object
}
\description{
cdpcx_hist
Plots a fancy cdpcx (extended chordal diatonic pitch class plot"
}
