% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{chorus_to_mcsv2}
\alias{chorus_to_mcsv2}
\title{chorus_to_mcsv2}
\usage{
chorus_to_mcsv2(chorus_tbl, tempo = 120, chorus_id = 1)
}
\arguments{
\item{chorus_tbl}{(data frame) generated phrases in data.frame}

\item{tempo}{(double scale) Tempo (bpm) of  the generated solo.}

\item{chorus_id}{(integer scalar) Id of chorus for coordination}
}
\value{
A solo data frame in MCSV2 format
}
\description{
This function formats a generated chorus to mcsv2-compatible data frame
}
