% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_int_pattern}
\alias{classify_int_pattern}
\title{classify_int_pattern
Classifies interval pattern in a main type (cool/uncool) and a subtype (scale, repetition, other, trill)}
\usage{
classify_int_pattern(pattern)
}
\arguments{
\item{pattern}{(string vector) Vector of interval patterns as comma-separated list of semi-tone intervals}
}
\value{
Data frame of classifications
}
\description{
classify_int_pattern
Classifies interval pattern in a main type (cool/uncool) and a subtype (scale, repetition, other, trill)
}
