% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{cpc_hist}
\alias{cpc_hist}
\title{cpc_hist}
\usage{
cpc_hist(data, id = NULL, percentage = T, fill_var = NULL, cpc_col = "cpc_raw")
}
\arguments{
\item{data}{(data frame) must contain "cpc_col" as variable with chordal pitch class data}

\item{id}{(string or integer) Optional set of ids to filter}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{fill_var}{(string) extra fill variable}

\item{cpc_col}{(string) column name where chordal pitch class data is stored, defaults to "cpc_raw"}
}
\value{
A ggplot2 object
}
\description{
Plots a fancy chordal pitch class plot
}
