% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{cpc_plot_solo}
\alias{cpc_plot_solo}
\title{cpc_plot_solo}
\usage{
cpc_plot_solo(solo, by_chord = T)
}
\arguments{
\item{solo}{(solo data frame) MCSV2 formatted solo data.frame}

\item{by_chord}{(logical scalar) Flag to add chord facets}
}
\description{
This function produces a bar plot of CPC values from a solo (MCSV2) data frame
}
