% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dir_or_warn}
\alias{create_dir_or_warn}
\title{Create directory or warn user if directory already exists}
\usage{
create_dir_or_warn(dir)
}
\arguments{
\item{dir}{Directory to attempt to create. User will be warned if the
directory already exists.}
}
\description{
Create directory or warn user if directory already exists
}
\keyword{internal}
