% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{create_leadsheet}
\alias{create_leadsheet}
\title{create_leadsheet}
\usage{
create_leadsheet(name, chords, length_beats)
}
\arguments{
\item{name}{(character scalar) Currently unused}

\item{chords}{(logical scalar) Flag if form shall be added}

\item{length_beats}{(integer scalar) length of chord in beats}
}
\value{
A leed sheet data frame
}
\description{
This function generates a lead sheet data frame from a vector of chord symbols and beat lengths per chord
}
