% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{create_leadsheet_from_chord_db}
\alias{create_leadsheet_from_chord_db}
\title{create_leadhseet_chord_db}
\usage{
create_leadsheet_from_chord_db(db, compid, name = NULL, with_form = F)
}
\arguments{
\item{db}{(chord data frame) Chord data frame to use}

\item{compid}{(integer or character scalar) If integer, then id of lead sheet in the chord data, if character, than title of song}

\item{name}{(character scalar) Currently unused}

\item{with_form}{(logical scalar) Flag if form shall be added}
}
\value{
A leed sheet data frame
}
\description{
This function generates a lead sheet data from parsed chord data frame (irb or wjd_chord_db),
}
