% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{create_leadsheet_from_irb}
\alias{create_leadsheet_from_irb}
\title{create_leadsheet_from_irb}
\usage{
create_leadsheet_from_irb(compid, name = NULL, with_form = F)
}
\arguments{
\item{compid}{(integer or character scalar) If integer, then id of lead sheet in the iRealBook data (data(irb)), if character, than title of song}

\item{name}{(character scalar) Currently unused}

\item{with_form}{(logical scalar) Flag if form shall be added}
}
\value{
A leed sheet data frame
}
\description{
This function generates a lead sheet data from parsed iRealBook data (irb),
}
