% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_csl}
\alias{download_csl}
\title{Download a CSL file from the CSL file repositories at either Zotero or
GitHub}
\usage{
download_csl(
  csl_style = "elsevier-harvard",
  csl_file = tempfile(fileext = ".csl"),
  zotero_first = TRUE
)
}
\arguments{
\item{csl_style}{A CSL style specification from the
\href{https://www.zotero.org/styles}{Zotero Style Repository} or \href{https://github.com/citation-style-language/styles}{GitHub CSL repository}. Should be the
\code{.csl} file name without the file extension, e.g. for \code{my-csl-style.csl},
provide \code{"my-csl-style"}. Default = \code{"elsevier-harvard"}.}

\item{csl_file}{A file path for \code{download_csl} to download to. Default =
\code{tempfile(fileext = ".csl")}.}

\item{zotero_first}{If \code{TRUE}, the function will attempt to download from the
Zotero Style Repository first. If \code{FALSE}, the function will attempt to
download from GitHub first. Default = \code{TRUE}.}
}
\value{
One-length character vector with the path of the downloaded \code{.csl}
file.
}
\description{
Citation Style Language (CSL) files are used to control referencing styles in
a range of softwares, including \href{https://github.com/jgm/citeproc}{citeproc},
the default citation processor in
\href{https://bookdown.org/yihui/bookdown/citations.html}{bookdown}.
\code{download_csl} downloads these for you from either the \href{https://www.zotero.org/styles}{Zotero Style Repository} or \href{https://github.com/citation-style-language/styles}{GitHub CSL repository}.
}
\examples{
\dontrun{
download_csl(csl_style = "elsevier_harvard")
}
}
