% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTL_interface.R
\name{DTL_similarity_search_results}
\alias{DTL_similarity_search_results}
\title{DTL_similarity_search_results
Runs a similarity search by the DTL similarity search API, same as DTL_similarity_search but returns data frame of results instead of result ID}
\usage{
DTL_similarity_search_results(
  search_patterns = "1,2,1,2,1,2,1,2",
  transformation = c("interval", "pitch", "fuzzyinterval"),
  database_names = "dtl,wjazzd,omnibook",
 
    metadata_filters = "{\\"dtl\\": {}, \\"wjazzd\\": {}, \\"esac\\": {}, \\"omnibook\\": {}}",
  filter_category = "0",
  minimum_similarity = 1,
  max_edit_distance = NA,
  max_length_difference = 0
)
}
\arguments{
\item{transformation}{(string) melodic representation to search in. Admissible values are "interval", "pitch", and "fuzzyinterval"}

\item{database_names}{(string) comma separated list of melodic database to search in. Database names are "dtl" (DTL1000 data set), "wjd" (Weimar Jazz Database), "omnibook" (Charlie Parker Omnibook), and "esac" (Essen Folks Song collection).}

\item{metadata_filters}{(string) Metadata filters for the databases, comma separated list of entries <dbname>: {<filter>}}

\item{filter_category}{(integer as string) I have no idea what this means}

\item{minimum_similarity}{(numeric) Value between 0 and 1 as the minimal similarity value between a pattern and the search pattern.}

\item{max_edit_distance}{(integer) Value of maximal edit distance between search pattern and data base patterns.}

\item{max_length_difference}{(integer) Maximum length difference between search pattern and result pattern}

\item{search_pattern}{(string vector) Array of search patterns, where a search pattern is a comma-separated list of search elements (depends on transformation)}
}
\value{
Search ID, to be used in DTL_get_results
}
\description{
DTL_similarity_search_results
Runs a similarity search by the DTL similarity search API, same as DTL_similarity_search but returns data frame of results instead of result ID
}
