% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{durclass_hist}
\alias{durclass_hist}
\title{durclass_hist
Plots a fancy duration class histogram}
\usage{
durclass_hist(
  data,
  id = NULL,
  id_var = "id",
  durclass_col = "durclass_abs_raw",
  durclass_col2 = "durclass_rel_raw",
  cmp_labels = c("Absolute", "Relative"),
  legend.x = 0.8,
  legend.y = 0.8,
  percentage = T,
  fill_var = NULL
)
}
\arguments{
\item{data}{(data frame) Must contain "fuzzyint_col" as column with duration clas data (integers from -2 (very short) to 2 (very long))}

\item{id}{(string or integer) Optional set of ids to filter}

\item{id_var}{(string or integer) Name of the column in data that contains the id var (allows grouping by different variables)}

\item{legend.x, }{legend.y (numeric) position of the legend, for nice inserts.}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{fill_var}{(string) Additional filler variable}

\item{fuzzyint_col}{(string) name of the column containing the interval data  (default "int_raw")}
}
\value{
ggplot2 object
}
\description{
durclass_hist
Plots a fancy duration class histogram
}
