% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esac_meta}
\alias{esac_meta}
\title{esac_meta}
\description{
Metadata for the  Essen Folk Song Database (EFSC) (\code{"id, cnr, collection, comment, esac_key, esac_title, function., melstring, region, signature, source, text, tunefamily, unit"})
}
