% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esac_transforms}
\alias{esac_transforms}
\title{esac_transforms}
\description{
Some basic raw features from the Essen Folk Song Database (EFSC) (\code{"id, int_raw, ioiclass_abs_raw, ioiclass_rel_raw, mcm_48"})
}
