% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simile.R
\name{find_best_transposition}
\alias{find_best_transposition}
\title{finds transposition that maximize raw edit distance of two pitch vectors
transposition in semitone of the *second* melody}
\usage{
find_best_transposition(pitch_vec1, pitch_vec2)
}
\description{
finds transposition that maximize raw edit distance of two pitch vectors
transposition in semitone of the *second* melody
}
