% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_grammar.R
\name{find_wba}
\alias{find_wba}
\title{find_wba}
\usage{
find_wba(x, debug = F)
}
\arguments{
\item{x}{(integer vector or character string) interval sequence}
}
\value{
Data frame with columns \code{length} (length of atom), \code{type} (WBA type),
\code{value} (interval string),
\code{direction} (direction of atom),
\code{start, end} (start and end position in sequence),
\code{pre_over, post_over} (amount of overlap to previous or following atom.)
}
\description{
This function parses a interval vector (or corresponding string) into its WBA constituents
}
