% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_grammar.R
\name{find_wba_by_phrase}
\alias{find_wba_by_phrase}
\title{find_wba_by_phrase}
\usage{
find_wba_by_phrase(data, id = NULL)
}
\arguments{
\item{data}{(data.frame) must contain int_raw, id and phrase_id columns,}

\item{id}{(string) string specifying id to use, if \code{NULL} the entire data frame is used.}
}
\value{
Data frame with columns \code{length} (length of atom), \code{type} (WBA type),
\code{value} (interval string),
\code{direction} (direction of atom),
\code{start, end} (start and end position in sequence),
\code{pre_over, post_over} (amount of overlap to previous or following atom.)
}
\description{
This function parses an \code{int_raw} column of  a data frame into its WBA constituents separately by phrases and optional by id.
 Columns \code{id} and \code{phrase_id} must be present.
}
