% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{fuzzyint_hist}
\alias{fuzzyint_hist}
\title{int_hist
Plots a fancy Fuzzy interval histogram}
\usage{
fuzzyint_hist(
  data,
  id = NULL,
  percentage = T,
  fill_var = NULL,
  fuzzyint_col = "fuzzyint_raw"
)
}
\arguments{
\item{data}{(data frame) Must contain "fuzzyint_col" as column with fuzzy interval data (integers -4 to 4)}

\item{id}{(string or integer) Optional set of ids to filter}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{fill_var}{(string) Additional filler variable}

\item{fuzzyint_col}{(string) name of the column containing the interval data  (default "int_raw")}
}
\value{
ggplot2 object
}
\description{
int_hist
Plots a fancy Fuzzy interval histogram
}
