% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solo_generator.R
\name{generate_chorus}
\alias{generate_chorus}
\title{generate_chorus}
\usage{
generate_chorus(
  lead_sheet,
  pitch_range = c(48, 84),
  tempo = 120,
  lick_to_line_ratio = 0.5,
  excludes = NULL
)
}
\arguments{
\item{lead_sheet}{(data frame) Lead sheet}

\item{pitch_range}{(Integer vector of length 2) Minimum and maximum MIDI pitch for solo generation.}

\item{tempo}{(double scale) Tempo (bpm) of  the generated solo.}

\item{lick_to_line_ratio}{(double scalar) Ratio of licks to lines}

\item{excludes}{(character vector) List of WBA atoms not to use during generation}
}
\value{
A solo data frame
}
\description{
This function generaties one chorus for a given lead sheet
}
