% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solo_generator.R
\name{generate_solo}
\alias{generate_solo}
\title{generate_solo}
\usage{
generate_solo(
  lead_sheet,
  n_chorus = 1,
  pitch_range = c(50, 83),
  tempo = 120,
  lick_to_line_ratio = 0.5,
  excludes = NULL
)
}
\arguments{
\item{lead_sheet}{(data frame) Lead sheet}

\item{n_chorus}{(Integer scalar) Number of choruses to generate.}

\item{pitch_range}{(Integer vector of length 2) Minimum and maximum MIDI pitch for solo generation.}

\item{tempo}{(double scale) Tempo (bpm) of  the generated solo.}

\item{lick_to_line_ratio}{(double scalar) Ratio of licks to lines}

\item{excludes}{(character vector) List of WBA atoms not to use during generation}
}
\value{
A solo data frame
}
\description{
This function generaties a solo for a given lead sheet
}
