% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ngrams_from_vector}
\alias{get_ngrams_from_vector}
\title{export}
\usage{
get_ngrams_from_vector(
  x,
  N,
  idx = NULL,
  ids = NULL,
  sep = ",",
  format = c("vector", "unique", "tibble")
)
}
\description{
export
}
