% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_analysis.R
\name{get_scale_degree_code}
\alias{get_scale_degree_code}
\title{get_scale_degree_code}
\usage{
get_scale_degree_code(chord_label, key, as_df = FALSE)
}
\arguments{
\item{chord_label}{(string) Valid chord symbol}

\item{key}{(string) valid key symbol}

\item{as_df}{(boolean) flag whether to return a data.frame or string vector}
}
\value{

}
\description{
get_scale_degree_code
}
