% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_stack.R
\name{get_value_context}
\alias{get_value_context}
\title{get_value_context
Retrieves all context n-grams for a value in a bigram stack}
\usage{
get_value_context(bs, value, pre = 0, post = 0, only_values = F)
}
\arguments{
\item{bs}{A bigram_stack object}

\item{value}{(string) value in bigram stack}

\item{pre}{(integer) Length of context before value}

\item{post}{(integer) Length of context after value}

\item{only_values}{(boolean) Flag, whether to return only unique context values}
}
\value{
Either a subset with context n-grams of the bigram stack or string of context values
}
\description{
get_value_context
Retrieves all context n-grams for a value in a bigram stack
}
