% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggassoc.R
\name{ggassoc}
\alias{ggassoc}
\title{ggassoc
Plots chisquare residuals for 2 x N tables, similar to vcd::asso}
\usage{
ggassoc(
  cont_tab,
  x_lab = "",
  y_lab = "",
  subtitle = "",
  spread = 2.25,
  text_size = 3,
  fill_colors = c("black", "red")
)
}
\arguments{
\item{cont_tab}{(table) 2 x N Contingency table}

\item{x_lab}{(string) Label for x-axis}

\item{y_lab}{(string) Label for y-axis}

\item{subtitle}{(string) subtitle for plot}

\item{spread}{(numeric) Spread factor for boxes}

\item{text_size}{(integer) Text size}

\item{fill_colors}{(string) 2-element vector of color values for filling boxes, defaults to black and red}
}
\value{
ggplot2
}
\description{
ggassoc
Plots chisquare residuals for 2 x N tables, similar to vcd::asso
}
