% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggassoc.R
\name{ggassoc2}
\alias{ggassoc2}
\title{ggassoc2
Plots chisquare residuals for 2 x N tables, similar to vcd::asso though width is proportional to overall frequencie, not expected values}
\usage{
ggassoc2(
  cont_tab,
  x_lab = "",
  y_lab = "",
  title = "Chi-Square Residuals",
  subtitle = "",
  layout = c("vertical", "horizontal"),
  fill_colors = jazzomat_palette$set1[1:2]
)
}
\arguments{
\item{cont_tab}{(table) 2 x N Contingency table}

\item{x_lab}{(string) Label for x-axis}

\item{y_lab}{(string) Label for y-axis}

\item{subtitle}{(string) subtitle for plot}

\item{fill_colors}{(string) 2-element vector of color values for filling boxes, defaults to black and red}

\item{layput}{(string)  Requests either  "vertical" (default) or "horizontal" layout of categorial residuals}
}
\value{
ggplot2
}
\description{
ggassoc2
Plots chisquare residuals for 2 x N tables, similar to vcd::asso though width is proportional to overall frequencie, not expected values
}
