% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{html-formatting}
\alias{html-formatting}
\alias{html_bold}
\alias{html_italic}
\alias{html_superscript}
\alias{html_subscript}
\alias{html_strikethrough}
\title{Add basic HTML tags to strings}
\usage{
html_bold(str)

html_italic(str)

html_superscript(str)

html_subscript(str)

html_strikethrough(str)
}
\arguments{
\item{str}{A string or vector of strings to be tagged.}
}
\value{
A character vector of the same length as \code{str}.
}
\description{
These functions take a vector of input strings, and tags appropriate to each
function. I find these most useful for inline code in R Markdown documents,
where I'd find manual HTML tagging more error prone.
}
\section{Functions}{
\itemize{
\item \code{html_bold()}: Add HTML strong (\if{html}{\out{<strong>}}bold\if{html}{\out{</strong>}}) tag

\item \code{html_italic()}: Add HTML emphasis (\if{html}{\out{<em>}}italic\if{html}{\out{</em>}}) tag

\item \code{html_superscript()}: Add HTML \if{html}{\out{<sup>}}superscript\if{html}{\out{</sup>}} tag

\item \code{html_subscript()}: Add HTML \if{html}{\out{<sub>}}subscript\if{html}{\out{</sub>}} tag

\item \code{html_strikethrough()}: Add HTML \if{html}{\out{<del>}}strikethrough\if{html}{\out{</del>}} tag

}}
\examples{
# Add a tag for bold
html_bold(str = "Text to make bold")

# Add a tag for italics
html_italic(str = "Text to make italicised")

# Apply tags to a vector of strings
html_subscript(str = c("all", "small"))
}
\seealso{
Other string formatting functions for R Markdown: 
\code{\link{latex-formatting}},
\code{\link{md-formatting}}
}
\concept{string formatting functions for R Markdown}
