% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{html_bold}
\alias{html_bold}
\alias{html_italic}
\alias{html_superscript}
\alias{html_subscript}
\alias{html_strikethrough}
\title{Add basic HTML tags to strings}
\usage{
html_bold(str)

html_italic(str)

html_superscript(str)

html_subscript(str)

html_strikethrough(str)
}
\arguments{
\item{str}{A string or vector of strings to be tagged.}
}
\value{
A character vector with HTML tags applied, the same length as \code{str}.
}
\description{
These functions take a vector of input strings, and apply the
approriate tag for each function.
}
\examples{
# Add a tag for bold
html_bold(str = "Text to make bold")

# Add a tag for italics
html_italic(str = "Text to make italicised")

# Apply tags to a vector of strings
html_subscript(str = c("all", "small"))
}
