% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{instrument_ranges}
\alias{instrument_ranges}
\title{instrument_ranges}
\description{
Tibble of instrument ranges, extracted from the Weimar Jazz Database, min(imum), max(imum),  5% (q5) and 95% (q95) percentiles
}
