% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{int_hist}
\alias{int_hist}
\title{int_hist
Plots a fancy Semitone interval histogram}
\usage{
int_hist(data, id = NULL, cut_off = 25, percentage = T, int_col = "int_raw")
}
\arguments{
\item{data}{(data frame) Must contain "int_col" as column with MCM data}

\item{id}{(string or integer) Optional set of ids to filter}

\item{cut_off}{(integer) Cutoff the distribution beyond this value (default 25)}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{int_col}{(string) name of the column containing the interval data  (default "int_raw")}
}
\value{
ggplot2 object
}
\description{
int_hist
Plots a fancy Semitone interval histogram
}
