% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{latex-formatting}
\alias{latex-formatting}
\alias{tex_bold}
\alias{tex_italic}
\alias{tex_code}
\alias{tex_superscript}
\title{Apply basic LaTeX formatting commands to strings}
\usage{
tex_bold(str)

tex_italic(str)

tex_code(str)

tex_superscript(str)
}
\arguments{
\item{str}{A string or vector of strings to be tagged.}
}
\value{
A character vector of the same length as \code{str}.
}
\description{
These functions take a vector of input strings, and insert these into
appropriate LaTeX commands for each function. I have found these useful when
preparing PDFs with R Markdown, especially in tasks like making certain
columns appear bold within LaTeX table output.
}
\section{Functions}{
\itemize{
\item \code{tex_bold()}: Add a LaTex formatting command for bold text
to a string

\item \code{tex_italic()}: Add a LaTex formatting command for italicised
text to a string

\item \code{tex_code()}: Add a LaTex formatting command for
teletype/typewriter/monospace text to a string

\item \code{tex_superscript()}: Add a LaTex formatting command for
superscripted text to a string

}}
\examples{
# Add command for bold
tex_bold(str = "Text to make bold")

# Add command for italics
tex_italic(str = "Text to make italicised")

# Apply command to a vector of strings
tex_code(str = c("As", "typewriter"))
}
\seealso{
Other string formatting functions for R Markdown: 
\code{\link{html-formatting}},
\code{\link{md-formatting}}
}
\concept{string formatting functions for R Markdown}
