% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solo_generator.R
\name{make_many_solos}
\alias{make_many_solos}
\title{make_many_solos}
\usage{
make_many_solos(
  n,
  fname = "solos/solo",
  lead_sheet = sologenerator::F_blues,
  lick_to_line_ratio = 0.7,
  excludes = NULL
)
}
\arguments{
\item{n}{(Integer scalar) Number of solos to generate.}

\item{fname}{(Scalar character) path for saving solo files, used as a base name}

\item{lead_sheet}{(data frame) Lead sheet}

\item{lick_to_line_ratio}{(scalar double) Ratio of licks to lines}

\item{excludes}{(character vector) List of WBA atoms not to use during generation}
}
\description{
This function generaties many one chorus solos from a given lead sheets
}
