% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{mcm_hist}
\alias{mcm_hist}
\title{mcm_hist}
\usage{
mcm_hist(
  data,
  id = NULL,
  percentage = T,
  fill_var = NULL,
  mcm48_col = "mcm_48",
  x_labels = labels[["metric_pos_4_4"]]
)
}
\arguments{
\item{data}{(data frame) Must contain "mcm48_col" as column with MCM data}

\item{id}{(string or integer) Optional set of ids to filter}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{fill_var}{(string) extra fill variable}

\item{mcm48_col}{(string) column name where MCM 48 data is stored, defaults to "mcm_48"}

\item{x_labels}{(string vector) Labels for x axis, defaults to 4/4 labels}
}
\value{
gg2plot object
}
\description{
Plots a fancy Metrical Circle Map (N = 48)  plot
}
