% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{md-formatting}
\alias{md-formatting}
\alias{md_bold}
\alias{md_italic}
\alias{md_code}
\alias{md_link}
\alias{md_superscript}
\alias{md_subscript}
\alias{md_strikethrough}
\alias{md_math}
\alias{md_eq}
\alias{md_footnote}
\title{Add basic Markdown formatting to strings}
\usage{
md_bold(str)

md_italic(str)

md_code(str)

md_link(str, url)

md_superscript(str)

md_subscript(str)

md_strikethrough(str)

md_math(str)

md_eq(str)

md_footnote(str)
}
\arguments{
\item{str}{A string or vector of strings to be tagged.}

\item{url}{URL to use in eventual hyperlink}
}
\value{
A character vector of the same length as \code{str}.
}
\description{
These functions take a vector of input strings, and apply approriate Markdown
formatting for each function. I often find then useful when making tables
with the \href{https://gt.rstudio.com/}{gt} package. Technically these functions
utilise R Markdown/pandoc-flavoured Markdown, but that's getting into the
weeds a bit.
}
\section{Functions}{
\itemize{
\item \code{md_bold()}: Format \strong{bold} Markdown text

\item \code{md_italic()}: Format \emph{italic} Markdown text

\item \code{md_code()}: Format in-line Markdown \code{code}

\item \code{md_link()}: Format text for a Markdown
\href{https://roxygen2.r-lib.org/}{hyperlink}

\item \code{md_superscript()}: Format superscripted Markdown text

\item \code{md_subscript()}: Format subscripted Markdown text

\item \code{md_strikethrough()}: Format struck-through Markdown text

\item \code{md_math()}: Format in-line Markdown mathematics

\item \code{md_eq()}: Format Markdown equations with their own paragraph

\item \code{md_footnote()}: Format text for a Markdown footnote

}}
\examples{
# Add a tag for bold
md_bold(str = "Text to make bold")

# Add a tag for italics
md_italic(str = "Text to make italicised")

# Apply tags to a vector of strings
md_code(str = c("As", "code"))
}
\seealso{
Other string formatting functions for R Markdown: 
\code{\link{html-formatting}},
\code{\link{latex-formatting}}
}
\concept{string formatting functions for R Markdown}
