% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_grammar.R
\name{messagef}
\alias{messagef}
\title{messagef
Combines sprintf and message}
\usage{
messagef(...)
}
\arguments{
\item{...}{Arguments to be passed to message()}
}
\description{
messagef
Combines sprintf and message
}
