% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{nice_cdpcx}
\alias{nice_cdpcx}
\title{nice_cdpcx
Makes a nice (sorted, labeled) cdpcx factor}
\usage{
nice_cdpcx(cdpcx_factor)
}
\arguments{
\item{cdpcx_factor}{}
}
\value{
factor of vector
}
\description{
nice_cdpcx
Makes a nice (sorted, labeled) cdpcx factor
}
