% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omnibook_transforms}
\alias{omnibook_transforms}
\title{omnibook_transforms}
\description{
Some basic raw features from the Weimar Jazz Database (WJD) (\code{"id, beat1, beat3, beatall, phrasbeg, phrasend, cdpcx_raw_all, cpc_raw_all, durclass_abs_raw, durclass_rel_raw, metrical_position, chord_types_raw, chords_raw, phrase_id_raw, mcm_48, pitch_raw, fuzzyint_raw, int_raw, ioiclass_abs_raw, ioiclass_rel_raw, pos"})
}
