% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkR-package.R
\docType{package}
\name{parkR-package}
\alias{parkR}
\alias{parkR-package}
\title{parkR: Generating monophonic jazz solos for chord sequences and more}
\description{
Generates jazz solos using Weimar Bebop Alphabet atoms and midlevel units based on the Weimar Jazz Database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/klausfrieler/parkR}
  \item Report bugs at \url{https://github.com/klausfrieler/parkR/issues}
}

}
\author{
\strong{Maintainer}: Klaus Frieler \email{klaus.frieler@ae.mpg.de}

}
\keyword{internal}
