% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{parkr_options}
\alias{parkr_options}
\title{Options}
\usage{
parkr_options(...)
}
\arguments{
\item{...}{a list of options.}
}
\value{
The function prints all set options if called with no arguments.
When setting options, nothing is returned.
}
\description{
Options for parkR package.
}
\details{
Currently only \code{debug} are used.
}
\examples{
parkr_options()

parkr_options(debug = TRUE)
}
