% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{parse_chord}
\alias{parse_chord}
\title{parse_chord}
\usage{
parse_chord(chord_label)
}
\arguments{
\item{chord_label}{(character vector) chord labels to parse}
}
\value{
data frame with columns \code{root} (tone name), \code{type} (basic chord type), \code{pc} (pitch class of root),
\code{bass} (extra bass note by slash chord), \code{bass_pc} (pitch class of ),
\code{ext} (string of comma separated extensions)
}
\description{
This function parses a chord label into its constituents
}
