% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{pc_hist}
\alias{pc_hist}
\title{pc_hist}
\usage{
pc_hist(data, id = NULL, percentage = T, fill_var = NULL, pc_col = "pc_raw")
}
\arguments{
\item{data}{(data frame) must contain "pc_col" as variable with pitch class data}

\item{id}{(string or integer) Optional set of ids to filter}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{fill_var}{(string) extra fill variable}

\item{pc_col}{(string) column name where pitch class data is stored, defaults to "pc_raw"}
}
\value{
A ggplot2 object
}
\description{
Plots a fancy pitch class plot
}
