% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solo_generator.R
\name{phrase_to_mcsv2}
\alias{phrase_to_mcsv2}
\title{phrase_to_mcsv2}
\usage{
phrase_to_mcsv2(phrase_tbl, tempo = 120, phrase_id = 1, chorus_id = 1)
}
\arguments{
\item{phrase_tbl}{(data frame) generated phrase data.frame}

\item{tempo}{(double scale) Tempo (bpm) of  the generated solo.}

\item{phrase_id}{(integer scalar) Id of phrase for coordination}

\item{chorus_id}{(integer scalar) Id of chorus for coordination}
}
\value{
A solo data frame in MCSV2 format
}
\description{
This function format a generated phrase to mcsv2-like data fraem
}
