% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{piano_roll}
\alias{piano_roll}
\title{piano_roll}
\usage{
piano_roll(solo, by_chorus = T)
}
\arguments{
\item{solo}{(solo data frame) MCSV2 formatted solo data.frame}

\item{by_chorus}{(logical scalar) Flag to add chorus facets}
}
\description{
This function produces a piano roll of a colo
}
