% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{pitch_hist}
\alias{pitch_hist}
\title{pitch_hist}
\usage{
pitch_hist(
  data,
  id = NULL,
  reduced_labels = F,
  percentage = T,
  fill_var = NULL,
  pitch_col = "pitch_raw"
)
}
\arguments{
\item{data}{(data frame) must contain "pitch_col" as variable with (MIDI) pitch  data}

\item{id}{(string or integer) Optional set of ids to filter}

\item{reduced_labels}{(boolean) Flag, whether to use reduced pitch labels}

\item{percentage}{(boolean) Flag, whether to use percentage scale on y-axis}

\item{fill_var}{(string) extra fill variable}

\item{pitch_col}{(string) column name where pitch  data is stored, defaults to "pitch_raw"}
}
\value{
A ggplot2 object
}
\description{
Plots a fancy pitch  plot
}
