% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{plot_z_values}
\alias{plot_z_values}
\title{Needle pLot for a range of z-transformed variables}
\usage{
plot_z_values(
  data,
  facet_var = NULL,
  max_z = 4,
  point_size = 3,
  base_size = NULL,
  sort_by_value = F
)
}
\arguments{
\item{data}{(data frame) A data frame in long format, with a column "name" that contains the variable names to compare
for, and a column "value", that contains z-transformed values.}

\item{facet_var}{(string) Name of column containing a factor variable for facetting (default is NULL, i.e., no facets)}

\item{max_z}{(numeric) Max z-value to consider}

\item{point_size}{(numeric) size of the "needle heads"}
}
\value{
A ggplot object
}
\description{
Needle pLot for a range of z-transformed variables
}
