% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_grammar.R
\name{printf}
\alias{printf}
\title{printf
Combines sprintf and print}
\usage{
printf(...)
}
\arguments{
\item{...}{Arguments to be passed to message()}
}
\description{
printf
Combines sprintf and print
}
