% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_crossref.R
\name{rmd-crossref}
\alias{rmd-crossref}
\alias{rmd_cref_tab}
\alias{rmd_cref_fig}
\alias{rmd_cref_eq}
\title{Construct bookdown cross-references for tables, figures, or equations}
\usage{
rmd_cref_tab(label)

rmd_cref_fig(label)

rmd_cref_eq(label)
}
\arguments{
\item{label}{The chunk label for your table, figure, or equation.}
}
\description{
These functions each use a code chunk label to cross-reference, according to
the format specifed for bookdown \href{https://bookdown.org/yihui/rmarkdown-cookbook/cross-ref.html}{cross-referencing}.
}
\section{Functions}{
\itemize{
\item \code{rmd_cref_tab()}: Cross-reference a table in R Markdown documents

\item \code{rmd_cref_fig()}: Cross-reference a figure in R Markdown documents

\item \code{rmd_cref_eq()}: Cross-reference an equation in R Markdown documents

}}
\note{
This will only work in a bookdown-style output, like
\code{\link[bookdown:html_document2]{bookdown::html_document2()}} or \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}.
}
\examples{
rmd_cref_tab(label = "life-by-sex-table")

rmd_cref_fig(label = "life-by-sex-figure")

rmd_cref_eq(label = "life-by-sex-equation")
}
