% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_crossref.R
\name{rmd_cref_tab}
\alias{rmd_cref_tab}
\alias{rmd_cref_fig}
\alias{rmd_cref_eq}
\title{Construct bookdown cross-references for tables, figures, or equations}
\usage{
rmd_cref_tab(label)

rmd_cref_fig(label)

rmd_cref_eq(label)
}
\arguments{
\item{label}{The chunk label for your figure, table or equation to
cross-reference.}
}
\description{
Construct bookdown cross-references for tables, figures, or equations
}
\note{
This will only work in a \code{{bookdown}} document, such as those produced
by \code{use_rmd_html()}.
}
\examples{
rmd_cref_tab(label = "life-by-sex-table")
rmd_cref_fig(label = "life-by-sex-figure")
rmd_cref_eq(label = "life-by-sex-equation")
}
