% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_crossref.R
\name{rmd_crossref}
\alias{rmd_crossref}
\title{Format an R Markdown cross-reference for tables, figures, or equations}
\usage{
rmd_crossref(type = "", label)
}
\arguments{
\item{type}{The type of environment being referenced. Can be one of \code{"fig"},
\code{"tab"}, or \code{"eq"}.}

\item{label}{The chunk label for your figure, table or equation to
cross-reference.}
}
\description{
Format an R Markdown cross-reference for tables, figures, or equations
}
\keyword{internal}
