% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lines.R
\name{rmd_lines_child}
\alias{rmd_lines_child}
\title{Generate lines for a simple R Markdown child document}
\usage{
rmd_lines_child(child_number)
}
\arguments{
\item{child_number}{Which section is this? Determines header and body text.}
}
\description{
Generate lines for a simple R Markdown child document
}
\keyword{internal}
