% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lines.R
\name{rmd_lines_citation_reminder}
\alias{rmd_lines_citation_reminder}
\title{Generate lines which describe simple how to cite in R Markdown}
\usage{
rmd_lines_citation_reminder()
}
\value{
2-length character vector describing how to cite in R Markdown.
}
\description{
Generate lines which describe simple how to cite in R Markdown
}
\keyword{internal}
