% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lines.R
\name{rmd_lines_parent}
\alias{rmd_lines_parent}
\title{Generate lines for an R Markdown document with children}
\usage{
rmd_lines_parent(
  title,
  author,
  date = Sys.Date(),
  bib_path,
  csl_path,
  child_dir,
  child_prefix
)
}
\arguments{
\item{title}{Document title.}

\item{author}{Document author.}

\item{date}{Document date. Default = \code{Sys.Date()}, which will update whenever
the document is knitted.}

\item{bib_path}{The file path for the bibliography file.}

\item{csl_path}{The file path for the CSL file used to generate citations.}
}
\description{
Generate lines for an R Markdown document with children
}
\keyword{internal}
