% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lines.R
\name{rmd_lines_setup}
\alias{rmd_lines_setup}
\title{Generate lines for a basic R Markdown setup code block}
\usage{
rmd_lines_setup()
}
\description{
Generate lines for a basic R Markdown setup code block
}
\keyword{internal}
