% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lines.R
\name{rmd_lines_yaml}
\alias{rmd_lines_yaml}
\title{Generate lines for YAML intro to an R Markdown document}
\usage{
rmd_lines_yaml(
  title = "",
  author = "",
  date = Sys.Date(),
  bib_path = "bibliography/references.bib",
  csl_path = "bibliography/elsevier-harvard.csl"
)
}
\arguments{
\item{title}{Document title.}

\item{author}{Document author.}

\item{date}{Document date. Default = \code{Sys.Date()}, which will update whenever
the document is knitted.}

\item{bib_path}{The file path for the bibliography file.}

\item{csl_path}{The file path for the CSL file used to generate citations.}
}
\value{
Character vector with lines corresponding to YAML heading for R
Markdown documents.
}
\description{
Generate lines for YAML intro to an R Markdown document
}
\keyword{internal}
