% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{simulate_wjd}
\alias{simulate_wjd}
\title{simulate_wjd}
\usage{
simulate_wjd(ids = 1:456)
}
\arguments{
\item{ids}{(integer vector) List of ids to create, ids shoudl be in the range 1 to 456, everything else will be ignored}
}
\value{
A data frame of generated solos
}
\description{
This function simulates a full Weimar Jazz Database, based on lead sheet, chorus counts and tempo
}
