% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{solo_to_lilypond}
\alias{solo_to_lilypond}
\title{solo_to_mcsv2}
\usage{
solo_to_lilypond(
  solo_tbl,
  file = "solo.pdf",
  key = "c",
  tempo = 120,
  lead_sheet = NULL,
  ...
)
}
\arguments{
\item{solo_tbl}{(data frame) generated phrases in data.frame}

\item{tempo}{(double scale) Tempo (bpm) of  the generated solo.}
}
\value{
A solo data frame in MCSV2 format
}
\description{
This function formats a generated solo to mcsv2-compatible data fraem
}
