% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solo_generator.R
\name{solo_to_mcsv2}
\alias{solo_to_mcsv2}
\title{solo_to_mcsv2}
\usage{
solo_to_mcsv2(solo_tbl, tempo = 120)
}
\arguments{
\item{tempo}{(double scale) Tempo (bpm) of  the generated solo.}

\item{chorus_tbl}{(data frame) generated phrase data.frame}
}
\value{
A solo data frame in MCSV2 format
}
\description{
This function formats a generated solo to mcsv2-compatible data fraem
}
