% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{tex_bold}
\alias{tex_bold}
\alias{tex_italic}
\alias{tex_superscript}
\alias{tex_code}
\title{Apply basic LaTeX formatting commands to strings}
\usage{
tex_bold(str)

tex_italic(str)

tex_superscript(str)

tex_code(str)
}
\arguments{
\item{str}{A string or vector of strings to be tagged.}
}
\value{
A character vector with HTML tags applied, the same length as \code{str}.
}
\description{
These functions take a vector of input strings, and apply the
approriate LaTeX commands for each function. I have found these useful when
preparing PDFs with R Markdown.
}
\examples{
# Add command for bold
tex_bold(str = "Text to make bold")

# Add command for italics
tex_italic(str = "Text to make italicised")

# Apply command to a vector of strings
tex_code(str = c("all", "small"))
}
