% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rmarkdown.R
\name{use_bibliography}
\alias{use_bibliography}
\title{Generate a directory with a bibliography and CSL file.}
\usage{
use_bibliography(
  bib_dir = "bibliography",
  bib_file = "references.bib",
  csl_style = "elsevier-harvard",
  overwrite = FALSE
)
}
\arguments{
\item{bib_dir}{Name of new directory.}

\item{bib_file}{Name of new \code{.bib} file.}

\item{csl_style}{A CSL style specification from the
\href{https://www.zotero.org/styles}{Zotero Style Repository} or \href{https://github.com/citation-style-language/styles}{GitHub CSL repository}. Should be the
\code{.csl} file name without the file extension, e.g. for \code{my-csl-style.csl},
provide "my-csl-style". Default = \code{"elsevier-harvard"}.}

\item{overwrite}{Should the function overwrite files without asking? Default
= \code{FALSE}.}
}
\description{
Generate a directory with a bibliography and CSL file.
}
