% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{use_hexsticker}
\alias{use_hexsticker}
\title{Add or modify files for using an R package hex sticker}
\usage{
use_hexsticker(url = get_pkg_url())
}
\description{
Add or modify files for using an R package hex sticker
}
\keyword{internal}
