% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rmarkdown.R
\name{use_rmd_children}
\alias{use_rmd_children}
\title{Generate some R Markdown child documents in a specific folder}
\usage{
use_rmd_children(
  child_dir = "sections",
  n = 2,
  child_prefix = "sec",
  overwrite = FALSE
)
}
\arguments{
\item{child_dir}{Directory in which to write children \code{.rmd} documents.
Default = \code{"sections"}.}

\item{n}{Number of child documents to generate. Default = \code{2}.}

\item{child_prefix}{Prefix for child documents. Default = "sec".}

\item{overwrite}{If TRUE, do not ask before overwriting existing files.
Default = FALSE.}
}
\description{
Generate some R Markdown child documents in a specific folder
}
