% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rmarkdown.R
\name{use_rmd_html}
\alias{use_rmd_html}
\title{Setup an R Markdown files which will knit to HTML with my preferred defaults}
\usage{
use_rmd_html(
  rmd_file = "untitled.rmd",
  bib_dir = "bibliography",
  bib_file = "references.bib",
  csl_style = "elsevier-harvard",
  child_dir = NULL,
  child_prefix = "sec",
  overwrite = FALSE
)
}
\arguments{
\item{rmd_file}{Name of the main \code{.rmd} document, which will have YAML for
a \code{\link[bookdown:html_document2]{bookdown::html_document2()}} output. If \code{child_dir} is specified, this file
will contain a code chunk for including child \code{.rmd} files.}

\item{bib_dir}{Name of directory which will contain \code{.bib} and \code{.csl} file
for referencing. Default = \code{"bibliography"}.}

\item{bib_file}{File name for \code{.bib} file in \code{bib_dir}. Default =
\code{"references.bib"}.}

\item{csl_style}{A CSL style specification from the
\href{https://www.zotero.org/styles}{Zotero Style Repository} or \href{https://github.com/citation-style-language/styles}{GitHub CSL repository}. Should be the
\code{.csl} file name without the file extension, e.g. for \code{my-csl-style.csl},
provide "my-csl-style". Default = \code{"elsevier-harvard"}.}

\item{child_dir}{If not \code{NULL}, specifies the name of the directory which
will contain child \code{.rmd} documents, which will be created by
\code{use_rmd_sections()}. Default = \code{NULL}.}

\item{child_prefix}{The file name prefix for child \code{.rmd} files in the folder
specified by \code{child_dir}. Default = "sec".}

\item{overwrite}{Logical value denoting whether files should be overwritten
automatically. If \code{FALSE}, the function will ask before overwriting. If
\code{TRUE}, the function will ask, then overwrite. Default = FALSE.}
}
\description{
Sets up folders/files in the working directory for an R Markdown
document which will knit to HTML.
}
