% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wjd_transforms}
\alias{wjd_transforms}
\title{wjd_transforms}
\description{
Some basic raw features from the Weimar Jazz Database (WJD) (\code{"id, beat13, beatall, phrasbeg, phrasend, cdpcx_raw_all, cpc_raw_all, chord_types_raw, chords_raw, phrase_id_raw, mcm_48, pitch_raw, fuzzyint_raw, int_raw, ioiclass_abs_raw, ioiclass_rel_raw, parsons_raw, style"})

Metadata for the  Weimar Jazz Database (WJD) solos (\code{"id, number_notes, avgtempo, chord_changes, chorus_count, composer, filename_sv, form, full_title, genre, harmony_template, instrument, key, label, lineup, mbzid, performer, recordbib, recordingdate, recordingyear, recordtitle, rhythmfeel, signature, solo_time, soloend, solopart, solostart, status, style, tempoclass, title, titleaddon, tonality_type"})
}
