% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_mcsv2}
\alias{write_mcsv2}
\title{write_mcsv2}
\usage{
write_mcsv2(data, fname)
}
\arguments{
\item{data}{(data frame) MCSV2 formatted solo data.frame}

\item{fname}{(character scale) path to write to.}
}
\description{
This function write a MCSV2 formated data fraem to a file
}
