% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_package}
\alias{is_package}
\title{Is a directory an R package?}
\usage{
is_package(base_path = usethis::proj_get())
}
\arguments{
\item{base_path}{Directory to test with
\code{rprojroot::find_package_root_file()}. Default = \code{usethis::proj_get()}.}
}
\value{
Logical value based on whether \code{base_path} was an R package.
}
\description{
Is a directory an R package?
}
\keyword{internal}
