% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_analysis.R
\name{key_analysis}
\alias{key_analysis}
\title{key_analysis}
\usage{
key_analysis(
  lead_sheet = NULL,
  chord_stream = NULL,
  with_ii_v_filter = T,
  use_chord_complexity = F,
  add_metadata = T
)
}
\arguments{
\item{lead_sheet}{(data frame) A lead sheet data frame or NULL, if NULL \code{chord_stream} is used}

\item{chord_stream}{(character string of chord symbols) A sequence of chords or NULL, if NULL \code{lead_sheet} is used, if both are NULL then an error is issued}

\item{with_ii_v_filter}{(logical scalar)  Use ii-V detection (default is \code{TRUE}).}

\item{add_metadata}{(logical scalar) Flag if (lead sheet) metadata (lead sheet info, analysis stats, key stats, estimated overall key) shall be added to result as attributes (default is \code{TRUE}).}
}
\value{
A key analysis data frame
}
\description{
This function parses the \code{chord} column from a lead sheet data frame or a vector of chord labels
into local keys and corresponding
local scale degree annotations (using Roman numerals)
}
