% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{nice_cdpcx}
\alias{nice_cdpcx}
\title{nice_cdpcx
Makes a nice (sorted, labelled) CDPCX factor}
\usage{
nice_cdpcx(cdpcx_values)
}
\arguments{
\item{cdpcx_values}{A vector of CDPCX values}
}
\value{
vector of CDPCX labels (factor)
}
\description{
nice_cdpcx
Makes a nice (sorted, labelled) CDPCX factor
}
