% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sqglue}
\alias{sqglue}
\title{Construct a glue expression using square brackets}
\usage{
sqglue(...)
}
\arguments{
\item{...}{Unnamed arguments are taken to be expression string(s) to format.
Multiple inputs are concatenated together before formatting.
Named arguments are taken to be temporary variables available for
substitution.}
}
\value{
A one-length character vector.
}
\description{
Wrapper around \code{\link[glue:glue]{glue::glue()}}, that uses square brackets instead
of the usual curly braces. Can be useful when substituting R expressions in
strings for LaTeX code, where curly braces are often used for formatting.
}
\examples{
italics_str <- "italic text"
# Double-curly braces escape each other, so italics_str is not interpreted
glue::glue("I want this in \\\\textit{{italic_str}}!")

# Square braces, so italics_str is incorporated properly
sqglue("I want this in \\\\textit{[italics_str]}!")
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}, which this function wraps.
}
